2022 day 11 - with help
Had to get help from `/r/adventofcode` on this one.

The code was taking way to long to run because the numbers
were so huge. Doing some modulus trick, that I don't fully
understand, kept the numbers small enough for the code to run
sub-second.

for/while loops, pop'ing, list comprehensions, eval vs lambdas
did not have an impact on the runtime.

Keeping the numbers small was the secret sauce.

Good on ya, AoC!